payment_backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
decouple==0.0.7
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.39.1
django-background-task==0.1.8
django-cors-headers==3.5.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-debug-toolbar==1.10.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
filelock==3.0.12
idna==2.8
importlib-metadata==2.0.0
oauthlib==3.0.1
pep8==1.7.1
pycodestyle==2.5.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
razorpay==1.2.0
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.24.2
zipp==3.3.0
